Constants

Constants contains a number of constant arrays and dictionaries used to define which keys and attributes are expected to be present in a GRIB message, how to map grid types to correct representations of the data, etc...

These constants are shown in the documentation for ease of use and reference purposes.

CfGRIB.ALL_HEADER_DIMS — Constant

SubString{String}["13-element", "Array{String,1}:", "\"number\"", "\"level\"", "\"dataDate\"", "\"dataTime\"", "\"endStep\"", "\"time\"", "\"step\"", "\"valid\time\"", "\"verifying\time\"", "\"indexing\_time\"", "\"forecastMonth\"", "\"directionNumber\"", "\"frequencyNumber\""]

CfGRIB.ALL_KEYS — Constant

SubString{String}["60-element", "Array{String,1}:", "\"DxInMetres\"", "\"DyInMetres\"", "\"J\"", "\"K\"", "\"LaDInDegrees\"", "\"Latin1InDegrees\"", "\"Latin2InDegrees\"", "\"LoVInDegrees\"", "\"M\"", "\"N\"", "\"NV\"", "\"Nx\"", "\"Ny\"", "\"angleOfRotationInDegrees\"", "\"centre\"", "\"centreDescription\"", "\"cfName\"", "\"cfVarName\"", "\"dataDate\"", "\"dataTime\"", "\"dataType\"", "\"directionNumber\"", "\"edition\"", "\"endStep\"", "\"forecastMonth\"", "\"frequencyNumber\"", "\"gridDefinitionDescription\"", "\"gridType\"", "\"iDirectionIncrementInDegrees\"", "\"iScansNegatively\"", "\"indexing\time\"", "\"jDirectionIncrementInDegrees\"", "\"jPointsAreConsecutive\"", "\"jScansPositively\"", "\"latitudeOfFirstGridPointInDegrees\"", "\"latitudeOfLastGridPointInDegrees\"", "\"latitudeOfSouthernPoleInDegrees\"", "\"level\"", "\"longitudeOfFirstGridPointInDegrees\"", "\"longitudeOfLastGridPointInDegrees\"", "\"longitudeOfSouthernPoleInDegrees\"", "\"missingValue\"", "\"name\"", "\"number\"", "\"numberOfDirections\"", "\"numberOfFrequencies\"", "\"numberOfPoints\"", "\"paramId\"", "\"pl\"", "\"shortName\"", "\"step\"", "\"stepType\"", "\"stepUnits\"", "\"subCentre\"", "\"time\"", "\"totalNumber\"", "\"typeOfLevel\"", "\"units\"", "\"valid\time\"", "\"verifying\_time\""]

CfGRIB.ALL_REF_TIME_KEYS — Constant

SubString{String}["6-element", "Array{String,1}:", "\"time\"", "\"step\"", "\"valid\time\"", "\"verifying\time\"", "\"indexing\_time\"", "\"forecastMonth\""]

CfGRIB.COORD_ATTRS — Constant

SubString{String}["Dict{String,Dict{String,String}}", "with", "15", "entries:", "\"number\"", "=>", "Dict(\"units\"=>\"1\",\"long\name\"=>\"ensemble", "member", "numerical", "id\",\"standard\name\"=>\"realization\")", "\"time\"", "=>", "Dict(\"units\"=>\"seconds", "since", "1970-01-01T00:00:00\",\"calendar\"=>\"proleptic\gregorian\",\"long\name\"=>\"initial", "time", "of", "forecast\",\"standard\name\"=>\"forecast\reference\time\")", "\"depthBelowLand\"", "=>", "Dict(\"units\"=>\"m\",\"long\name\"=>\"soil", "depth\",\"standard\name\"=>\"depth\",\"positive\"=>\"down\")", "\"hybrid\"", "=>", "Dict(\"units\"=>\"1\",\"long\name\"=>\"hybrid", "level\",\"standard\name\"=>\"atmosphere\hybrid\sigma\pressure\coordinate\",\"positive\"=>\"down\")", "\"depthBelowLandLayer\"", "=>", "Dict(\"units\"=>\"m\",\"long\name\"=>\"soil", "depth\",\"standard\name\"=>\"depth\",\"positive\"=>\"down\")", "\"isobaricLayer\"", "=>", "Dict(\"units\"=>\"Pa\",\"long\name\"=>\"pressure\",\"standard\name\"=>\"air\pressure\",\"positive\"=>\"down\")", "\"step\"", "=>", "Dict(\"units\"=>\"hours\",\"long\name\"=>\"time", "since", "forecast\reference\time\",\"standard\name\"=>\"forecast\period\")", "\"isobaricInhPa\"", "=>", "Dict(\"units\"=>\"hPa\",\"stored\direction\"=>\"decreasing\",\"long\name\"=>\"pressure\",\"standard\name\"=>\"air\pressure\",\"positive\"=>\"down\")", "\"isobaricInPa\"", "=>", "Dict(\"units\"=>\"Pa\",\"stored\direction\"=>\"decreasing\",\"long\name\"=>\"pressure\",\"standard\name\"=>\"air\pressure\",\"positive\"=>\"down\")", "\"valid\time\"", "=>", "Dict(\"units\"=>\"seconds", "since", "1970-01-01T00:00:00\",\"calendar\"=>\"proleptic\gregorian\",\"long\name\"=>\"time\",\"standard\name\"=>\"time\")", "\"indexing\time\"", "=>", "Dict(\"units\"=>\"seconds", "since", "1970-01-01T00:00:00\",\"calendar\"=>\"proleptic\gregorian\",\"long\name\"=>\"nominal", "initial", "time", "of", "forecast\",\"standard\name\"=>\"forecast\reference\time\")", "\"heightAboveGround\"", "=>", "Dict(\"units\"=>\"m\",\"long\name\"=>\"height", "above", "the", "surface\",\"standard\name\"=>\"height\",\"positive\"=>\"up\")", "\"latitude\"", "=>", "Dict(\"units\"=>\"degrees\north\",\"long\name\"=>\"latitude\",\"standard\name\"=>\"latitude\")", "\"longitude\"", "=>", "Dict(\"units\"=>\"degrees\east\",\"long\name\"=>\"longitude\",\"standard\name\"=>\"longitude\")", "\"verifying\time\"", "=>", "Dict(\"units\"=>\"seconds", "since", "1970-01-01T00:00:00\",\"calendar\"=>\"proleptic\gregorian\",\"long\name\"=>\"time\",\"standard\_name\"=>\"time\")"]

CfGRIB.DATA_ATTRIBUTES_KEYS — Constant

SubString{String}["18-element", "Array{String,1}:", "\"paramId\"", "\"shortName\"", "\"units\"", "\"name\"", "\"cfName\"", "\"cfVarName\"", "\"dataType\"", "\"missingValue\"", "\"numberOfPoints\"", "\"totalNumber\"", "\"numberOfDirections\"", "\"numberOfFrequencies\"", "\"typeOfLevel\"", "\"NV\"", "\"stepUnits\"", "\"stepType\"", "\"gridType\"", "\"gridDefinitionDescription\""]

CfGRIB.DATA_TIME_KEYS — Constant

SubString{String}["3-element", "Array{String,1}:", "\"dataDate\"", "\"dataTime\"", "\"endStep\""]

CfGRIB.ENSEMBLE_KEYS — Constant

SubString{String}["1-element", "Array{String,1}:", "\"number\""]

CfGRIB.GLOBAL_ATTRIBUTES_KEYS — Constant

SubString{String}["4-element", "Array{String,1}:", "\"edition\"", "\"centre\"", "\"centreDescription\"", "\"subCentre\""]

CfGRIB.GRID_TYPES_2D_NON_DIMENSION_COORDS — Constant

SubString{String}["5-element", "Array{String,1}:", "\"rotated\ll\"", "\"rotated\gg\"", "\"lambert\"", "\"albers\"", "\"polar\_stereographic\""]

CfGRIB.GRID_TYPES_DIMENSION_COORDS — Constant

SubString{String}["2-element", "Array{String,1}:", "\"regular\ll\"", "\"regular\gg\""]

CfGRIB.GRID_TYPE_KEYS — Constant

SubString{String}["25-element", "Array{String,1}:", "\"Nx\"", "\"Ny\"", "\"angleOfRotationInDegrees\"", "\"iDirectionIncrementInDegrees\"", "\"iScansNegatively\"", "\"jPointsAreConsecutive\"", "\"jScansPositively\"", "\"latitudeOfFirstGridPointInDegrees\"", "\"latitudeOfLastGridPointInDegrees\"", "\"latitudeOfSouthernPoleInDegrees\"", "\"longitudeOfFirstGridPointInDegrees\"", "\"longitudeOfLastGridPointInDegrees\"", "\"longitudeOfSouthernPoleInDegrees\"", "\"N\"", "\"M\"", "\"K\"", "\"J\"", "\"jDirectionIncrementInDegrees\"", "\"LaDInDegrees\"", "\"LoVInDegrees\"", "\"DyInMetres\"", "\"DxInMetres\"", "\"Latin2InDegrees\"", "\"Latin1InDegrees\"", "\"pl\""]

CfGRIB.GRID_TYPE_MAP — Constant

SubString{String}["Dict{String,Array{String,1}}", "with", "8", "entries:", "\"rotated\gg\"", "=>", "[\"Nx\",", "\"Ny\",", "\"angleOfRotationInDegrees\",", "\"iDirectionIncrementInDegrees\",", "\"iScansNegatively\",", "\"jPointsAreConsecutive\",", "\"jScansPositively\",", "\"latitudeOfFirstGridPointInDegrees\",", "\"latitudeOfLastGridPointInDegrees\",", "\"latitudeOfSouthernPoleInDegrees\",", "\"longitudeOfFirstGridPointInDegrees\",", "\"longitudeOfLastGridPointInDegrees\",", "\"longitudeOfSouthernPoleInDegrees\",", "\"N\"]", "\"regular\gg\"", "=>", "[\"Nx\",", "\"iDirectionIncrementInDegrees\",", "\"iScansNegatively\",", "\"longitudeOfFirstGridPointInDegrees\",", "\"longitudeOfLastGridPointInDegrees\",", "\"N\",", "\"Ny\"]", "\"sh\"", "=>", "[\"M\",", "\"K\",", "\"J\"]", "\"regular\ll\"", "=>", "[\"Nx\",", "\"iDirectionIncrementInDegrees\",", "\"iScansNegatively\",", "\"longitudeOfFirstGridPointInDegrees\",", "\"longitudeOfLastGridPointInDegrees\",", "\"Ny\",", "\"jDirectionIncrementInDegrees\",", "\"jPointsAreConsecutive\",", "\"jScansPositively\",", "\"latitudeOfFirstGridPointInDegrees\",", "\"latitudeOfLastGridPointInDegrees\"]", "\"lambert\"", "=>", "[\"LaDInDegrees\",", "\"LoVInDegrees\",", "\"iScansNegatively\",", "\"jPointsAreConsecutive\",", "\"jScansPositively\",", "\"latitudeOfFirstGridPointInDegrees\",", "\"latitudeOfSouthernPoleInDegrees\",", "\"longitudeOfFirstGridPointInDegrees\",", "\"longitudeOfSouthernPoleInDegrees\",", "\"DyInMetres\",", "\"DxInMetres\",", "\"Latin2InDegrees\",", "\"Latin1InDegrees\",", "\"Ny\",", "\"Nx\"]", "\"reduced\gg\"", "=>", "[\"N\",", "\"pl\"]", "\"rotated\ll\"", "=>", "[\"Nx\",", "\"Ny\",", "\"angleOfRotationInDegrees\",", "\"iDirectionIncrementInDegrees\",", "\"iScansNegatively\",", "\"jDirectionIncrementInDegrees\",", "\"jPointsAreConsecutive\",", "\"jScansPositively\",", "\"latitudeOfFirstGridPointInDegrees\",", "\"latitudeOfLastGridPointInDegrees\",", "\"latitudeOfSouthernPoleInDegrees\",", "\"longitudeOfFirstGridPointInDegrees\",", "\"longitudeOfLastGridPointInDegrees\",", "\"longitudeOfSouthernPoleInDegrees\"]", "\"reduced\ll\"", "=>", "[\"Ny\",", "\"jDirectionIncrementInDegrees\",", "\"jPointsAreConsecutive\",", "\"jScansPositively\",", "\"latitudeOfFirstGridPointInDegrees\",", "\"latitudeOfLastGridPointInDegrees\"]"]

CfGRIB.SPECTRA_KEYS — Constant

SubString{String}["2-element", "Array{String,1}:", "\"directionNumber\"", "\"frequencyNumber\""]

CfGRIB.VERTICAL_KEYS — Constant

SubString{String}["1-element", "Array{String,1}:", "\"level\""]