Dataset

The dataset file contains the majority of the core code in CfGRIB.jl. It defines the key datatypes used throughout the package, as well as how data is read from the disk and how it gets interpreted.

CfGRIB.DataSet — Type

Summary

struct DataSet

Maps a GRIB file to the NetCDF Common Data Model with CF Conventions.

Fields

dimensions::OrderedCollections.OrderedDict{String,Int64}

OrderedDict{String,Int} of $DIMENSION_NAME => $DIMENSION_LENGTH.

variables::OrderedCollections.OrderedDict{String,CfGRIB.Variable}

OrderedDict{String,CfGRIB.Variable} of $DIMENSION_NAME => $DIMENSION_VARIABLE, where the the variable is a Variable.

attributes::OrderedCollections.OrderedDict{String,Any}

OrderedDict{String,Any} containing some metadata extracted from the file.

encoding::Dict{String,Any}

Dict{String,Any} containing metadata related to CfGRIB.jl, e.g. filter_by_keys

Constructors

DataSet(dimensions, variables, attributes, encoding)

defined at /home/runner/work/cfgrib.jl/cfgrib.jl/src/dataset.jl:136.

DataSet(path; encode_cf, errors, filter_by_keys, read_keys, squeeze, time_dims)

defined at /home/runner/work/cfgrib.jl/cfgrib.jl/src/dataset.jl:153.

CfGRIB.OnDiskArray — Type

Summary

struct OnDiskArray

Struct that contains metadata for an array, used to lazy-load the array from disk only when requested

Fields

grib_path::String

size::Tuple

offsets::OrderedCollections.OrderedDict

message_lengths::Array{Int64,1}

missing_value::Any

geo_ndim::Int64

dtype::Type

Constructors

OnDiskArray(grib_path, size, offsets, message_lengths, missing_value, geo_ndim, dtype)

defined at /home/runner/work/cfgrib.jl/cfgrib.jl/src/dataset.jl:27.