Dataset

The dataset file contains the majority of the core code in CfGRIB.jl. It defines the key datatypes used throughout the package, as well as how data is read from the disk and how it gets interpreted.

CfGRIB.DataSet — Type

Summary

struct DataSet

Map a GRIB file to the NetCDF Common Data Model with CF Conventions.

Fields

dimensions::OrderedCollections.OrderedDict

variables::OrderedCollections.OrderedDict

attributes::OrderedCollections.OrderedDict

encoding::Dict

Constructors

DataSet(dimensions, variables, attributes, encoding)
DataSet(dimensions, variables, attributes, encoding)

DataSet(path; read_keys, kwargs...)