CFMessage

CFMessage contains a number of utility functions for converting variables in a GRIB message to the values a user would expect to work with.

CfGRIB.COMPUTED_KEYS — Constant

COMPUTED_KEYS::Dict{String,Tuple{Function,Function}}

Dictionary which maps a key to a conversion method. The first function is the 'to' conversion, the second is 'from'.

TODO: Actually applying the from_grib_step function results in different values to cfgrib.py, so step -> (from_grib_step, to_grib_step) is currently disabled.

Currently converts:

    "time" => (from_grib_date_time, to_grib_date_time)

    "valid_time" => (
        message -> from_grib_date_time(message, date_key="validityDate", time_key="validityTime"),
        message -> to_grib_date_time(message, date_key="validityDate", time_key="validityTime"),
    )

    "verifying_time" => (from_grib_month, m -> throw(ErrorException("Unimplemented")))

    "indexing_time" => (
        message -> from_grib_date_time(message, date_key="indexingDate", time_key="indexingTime"),
        message -> to_grib_date_time(message, date_key="indexingDate", time_key="indexingTime"),
    )

Example

A GRIB message containing 20160501 as the date key and 0 as the time key would end up calling:

julia> CfGRIB.COMPUTED_KEYS["time"](20160501, 0)
1462060800

CfGRIB.DEFAULT_EPOCH — Constant

DateTime(1970, 1, 1, 0, 0)

Default epoch used for from_ and to_ methods, set to 1st of January 1970.

CfGRIB.GRIB_STEP_UNITS_TO_SECONDS — Constant

Union{Missing, Int64}[60, 3600, 86400, missing, missing, missing, missing, missing, missing, missing, 10800, 21600, 43200, 1, 900]

Array used to convert the grib step units to seconds. As Julia is 1-indexed, not 0 like Python, you should take care to correctly access the array (typically +1 to the step units).

CfGRIB.build_valid_time — Method

build_valid_time(
    time::Array{Int, 1}, step::Array{Int, 1}
)::Union{
    Tuple{Tuple{},Int64},
    Tuple{Tuple{String,String},Array{Int64,2}}
}

Returns a pair of (dims, data) based on the type of input, this function behaves like:

julia> CfGRIB.build_valid_time([10, 10], [10, 10])
(("time", "step"), [36010 36010; 36010 36010])

However if it receives two arrays of one element then it dispatches to build_valid_time(time::Int, step::Int):

julia> CfGRIB.build_valid_time([10], [10])
((), 36010)

CfGRIB.build_valid_time — Method

build_valid_time(
    time::Array{Int, 1}, step::Int
)::Tuple{Tuple{String},Array{Int64,1}}

Returns a pair of (dims, data) based on the type of input, this function behaves like:

julia> CfGRIB.build_valid_time([10], 10)
(("time",), [36010])

CfGRIB.build_valid_time — Method

build_valid_time(
    time::Int, step::Array{Int, 1}
)::Tuple{Tuple{String},Array{Int64,1}}

Returns a pair of (dims, data) based on the type of input, this function behaves like:

julia> CfGRIB.build_valid_time(1, [10])
(("step",), [36001])

CfGRIB.build_valid_time — Method

build_valid_time(
    time::Int, step::Int
)::Tuple{Tuple{},Int64}

Returns a pair of (dims, data) based on the type of input, this function behaves like:

julia> CfGRIB.build_valid_time(10, 10)
((), 36010)

CfGRIB.from_grib_date_time — Method

from_grib_date_time(
        message::GRIB.Message; date_key="dataDate",
        time_key="dataTime", epoch=DEFAULT_EPOCH
)

Pulls out the date and time from given keys and passes them to from_grib_date_time(::Int, ::Int)

CfGRIB.from_grib_date_time — Method

from_grib_date_time(date::Int, time::Int; epoch=DEFAULT_EPOCH)

Returns the integer seconds from epoch to the given date and time.

CfGRIB.from_grib_month — Function

from_grib_month(
    message::GRIB.Message,
    verifying_month_key="verifyingMonth",
    epoch=DEFAULT_EPOCH
)

Returns the integer seconds from the epoch to the verifying month value in the GRIB message.

CfGRIB.from_grib_step — Function

from_grib_step(
        message::GRIB.Message,
        step_key="endStep", step_unit_key="stepUnits"
)

Returns the step_key value in hours.

Uses GRIB_STEP_UNITS_TO_SECONDS to convert the step values to seconds, then divides by 3600.0 to get hours.

CfGRIB.read_message — Method

read_message(message::GRIB.Message, key::String)

Reads a specific key from a GRIB.jl message. Attempts to convert the raw value associated with that key using the COMPUTED_KEYS mapping to from_grib_* functions.