Constants

Constants contains a number of constant arrays and dictionaries used to define which keys and attributes are expected to be present in a GRIB message, how to map grid types to correct representations of the data, etc...

These constants are shown in the documentation for ease of use and reference purposes.

CfGRIB.ALL_HEADER_DIMS — Constant

["number", "level", "dataDate", "dataTime", "endStep", "time", "step", "valid_time", "verifying_time", "indexing_time", "forecastMonth", "directionNumber", "frequencyNumber"]

CfGRIB.ALL_KEYS — Constant

["DxInMetres", "DyInMetres", "J", "K", "LaDInDegrees", "Latin1InDegrees", "Latin2InDegrees", "LoVInDegrees", "M", "N", "NV", "Nx", "Ny", "angleOfRotationInDegrees", "centre", "centreDescription", "cfName", "cfVarName", "dataDate", "dataTime", "dataType", "directionNumber", "edition", "endStep", "forecastMonth", "frequencyNumber", "gridDefinitionDescription", "gridType", "iDirectionIncrementInDegrees", "iScansNegatively", "indexing_time", "jDirectionIncrementInDegrees", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees", "latitudeOfSouthernPoleInDegrees", "level", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "longitudeOfSouthernPoleInDegrees", "missingValue", "name", "number", "numberOfDirections", "numberOfFrequencies", "numberOfPoints", "paramId", "pl", "shortName", "step", "stepType", "stepUnits", "subCentre", "time", "totalNumber", "typeOfLevel", "units", "valid_time", "verifying_time"]

CfGRIB.ALL_REF_TIME_KEYS — Constant

["time", "step", "valid_time", "verifying_time", "indexing_time", "forecastMonth"]

CfGRIB.COORD_ATTRS — Constant

Dict("number" => Dict("units" => "1","long_name" => "ensemble member numerical id","standard_name" => "realization"),"time" => Dict("units" => "seconds since 1970-01-01T00:00:00","calendar" => "proleptic_gregorian","long_name" => "initial time of forecast","standard_name" => "forecast_reference_time"),"depthBelowLand" => Dict("units" => "m","long_name" => "soil depth","standard_name" => "depth","positive" => "down"),"hybrid" => Dict("units" => "1","long_name" => "hybrid level","standard_name" => "atmosphere_hybrid_sigma_pressure_coordinate","positive" => "down"),"depthBelowLandLayer" => Dict("units" => "m","long_name" => "soil depth","standard_name" => "depth","positive" => "down"),"isobaricLayer" => Dict("units" => "Pa","long_name" => "pressure","standard_name" => "air_pressure","positive" => "down"),"step" => Dict("units" => "hours","long_name" => "time since forecast_reference_time","standard_name" => "forecast_period"),"isobaricInhPa" => Dict("units" => "hPa","stored_direction" => "decreasing","long_name" => "pressure","standard_name" => "air_pressure","positive" => "down"),"isobaricInPa" => Dict("units" => "Pa","stored_direction" => "decreasing","long_name" => "pressure","standard_name" => "air_pressure","positive" => "down"),"valid_time" => Dict("units" => "seconds since 1970-01-01T00:00:00","calendar" => "proleptic_gregorian","long_name" => "time","standard_name" => "time"),"indexing_time" => Dict("units" => "seconds since 1970-01-01T00:00:00","calendar" => "proleptic_gregorian","long_name" => "nominal initial time of forecast","standard_name" => "forecast_reference_time"),"heightAboveGround" => Dict("units" => "m","long_name" => "height above the surface","standard_name" => "height","positive" => "up"),"latitude" => Dict("units" => "degrees_north","long_name" => "latitude","standard_name" => "latitude"),"longitude" => Dict("units" => "degrees_east","long_name" => "longitude","standard_name" => "longitude"),"verifying_time" => Dict("units" => "seconds since 1970-01-01T00:00:00","calendar" => "proleptic_gregorian","long_name" => "time","standard_name" => "time"))

CfGRIB.DATA_ATTRIBUTES_KEYS — Constant

["paramId", "shortName", "units", "name", "cfName", "cfVarName", "dataType", "missingValue", "numberOfPoints", "totalNumber", "numberOfDirections", "numberOfFrequencies", "typeOfLevel", "NV", "stepUnits", "stepType", "gridType", "gridDefinitionDescription"]

CfGRIB.DATA_TIME_KEYS — Constant

["dataDate", "dataTime", "endStep"]

CfGRIB.ENSEMBLE_KEYS — Constant

["number"]

CfGRIB.GLOBAL_ATTRIBUTES_KEYS — Constant

["edition", "centre", "centreDescription", "subCentre"]

CfGRIB.GRID_TYPES_2D_NON_DIMENSION_COORDS — Constant

["rotated_ll", "rotated_gg", "lambert", "albers", "polar_stereographic"]

CfGRIB.GRID_TYPES_DIMENSION_COORDS — Constant

["regular_ll", "regular_gg"]

CfGRIB.GRID_TYPE_KEYS — Constant

["Nx", "Ny", "angleOfRotationInDegrees", "iDirectionIncrementInDegrees", "iScansNegatively", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees", "latitudeOfSouthernPoleInDegrees", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "longitudeOfSouthernPoleInDegrees", "N", "M", "K", "J", "jDirectionIncrementInDegrees", "LaDInDegrees", "LoVInDegrees", "DyInMetres", "DxInMetres", "Latin2InDegrees", "Latin1InDegrees", "pl"]

CfGRIB.GRID_TYPE_MAP — Constant

Dict("rotated_gg" => ["Nx", "Ny", "angleOfRotationInDegrees", "iDirectionIncrementInDegrees", "iScansNegatively", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees", "latitudeOfSouthernPoleInDegrees", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "longitudeOfSouthernPoleInDegrees", "N"],"regular_gg" => ["Nx", "iDirectionIncrementInDegrees", "iScansNegatively", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "N", "Ny"],"sh" => ["M", "K", "J"],"regular_ll" => ["Nx", "iDirectionIncrementInDegrees", "iScansNegatively", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "Ny", "jDirectionIncrementInDegrees", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees"],"lambert" => ["LaDInDegrees", "LoVInDegrees", "iScansNegatively", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfSouthernPoleInDegrees", "longitudeOfFirstGridPointInDegrees", "longitudeOfSouthernPoleInDegrees", "DyInMetres", "DxInMetres", "Latin2InDegrees", "Latin1InDegrees", "Ny", "Nx"],"reduced_gg" => ["N", "pl"],"rotated_ll" => ["Nx", "Ny", "angleOfRotationInDegrees", "iDirectionIncrementInDegrees", "iScansNegatively", "jDirectionIncrementInDegrees", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees", "latitudeOfSouthernPoleInDegrees", "longitudeOfFirstGridPointInDegrees", "longitudeOfLastGridPointInDegrees", "longitudeOfSouthernPoleInDegrees"],"reduced_ll" => ["Ny", "jDirectionIncrementInDegrees", "jPointsAreConsecutive", "jScansPositively", "latitudeOfFirstGridPointInDegrees", "latitudeOfLastGridPointInDegrees"])

CfGRIB.SPECTRA_KEYS — Constant

["directionNumber", "frequencyNumber"]

CfGRIB.VERTICAL_KEYS — Constant

["level"]