Indexing

GRIB.jl is by CfGRIB to read the GRIB files, GRIB.jl provides its own ways to create, read, and filter an index, however cfgrib.py uses a different approach to indexing. So, CfGRIB.jl, being based on the python implementation, recreates the python approach instead of integrating with the GRIB.jl indexing approach.

The indexing file defines the FileIndex type, as well as the constructors for that type.

CfGRIB.FileIndex — Type

FileIndex

A mutable store for indices of a GRIB file

Constructors

FileIndex(grib_path::String, index_keys::Array{String, 1})

Fields

allowed_protocol_version::VersionNumber : Version number used when saving/hashing index files

grib_path::String : Path to the file the index belongs to

index_path::String : Path to the index cache file

index_keys::Array{String, 1} :

offsets::Array :

header_values::OrderedDict{String, Array} :

filter_by_keys::Dict :